Changed many double types to decimals
For the sake of avoiding floating point inaccuracy when dealing with currency calculations.

+++ Models/CropField.cs
@@ -4,7 +4,7 @@ public partial class CropField
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public double Hectares { get; set; }
+    public decimal Hectares { get; set; }
 
     public virtual ICollection<BalanceLedger> BalanceLedgers { get; set; } = new List<BalanceLedger>();
     public virtual ICollection<FieldEfficiency> FieldEfficiencies { get; set; } = new List<FieldEfficiency>();
